+++ src/ExpoYandexMapKitView.web.tsx
@@ -24,6 +24,9 @@ export const YandexMapView = React.forwardRef<YandexMapViewRef, YandexMapViewPro
         selectGeoObject: unsupported,
         deselectGeoObject: unsupported,
         fitToCoordinates: unsupported,
+        setIndoorLevel: unsupported,
+        addTileOverlay: unsupported,
+        removeTileOverlay: unsupported,
       };
     }, []);
     return null;